+++ src/tokens.js
@@ -2274,6 +2274,56 @@ module.exports = [
         ],
     },
     {
+        name: 'Scrim',
+        description: 'Values for transparent light and dark curtains that cover content.',
+        tokens: [
+            {
+                id: 'scrim-light-80',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: '$tp-scrim__light-80',
+                        value: 'rgba(255, 255, 255, 0.8)',
+                    },
+                    javascript: {
+                        name: 'tpScrimLight80',
+                        value: 'rgba(255, 255, 255, 0.8)',
+                    },
+                    ios: {
+                        name: 'light80',
+                        value: 'UIColor(red: 1.0, green: 1.0, blue: 1.0, alpha: 0.8)',
+                    },
+                    android: {
+                        name: 'tp_scrim_light_80',
+                        value: '#ccffffff',
+                    },
+                },
+            },
+            {
+                id: 'scrim-dark-80',
+                type: 'string',
+                platforms: {
+                    scss: {
+                        name: '$tp-scrim__dark-80',
+                        value: 'rgba(0, 0, 0, 0.8)',
+                    },
+                    javascript: {
+                        name: 'tpScrimDark80',
+                        value: 'rgba(0, 0, 0, 0.8)',
+                    },
+                    ios: {
+                        name: 'dark80',
+                        value: 'UIColor(red: 0.0, green: 0.0, blue: 0.0, alpha: 0.8)',
+                    },
+                    android: {
+                        name: 'tp_scrim_dark_80',
+                        value: '#cc000000',
+                    },
+                },
+            },
+        ],
+    },
+    {
         name: 'Space',
         tokens: [
             {